requirements.txt
streamlit==1.45.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain
langchain-core
langchain-community
langgraph
langgraph-prebuilt
typing-extensions==4.12.2
python-dateutil==2.9.0.post0
requests==2.32.3
numpy==1.26.4
pandas==2.2.1 
langchain-groq
sounddevice
librosa
watchdog
streamlit-mic-recorder
scipy
groq